backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
psycopg2-binary==2.9.3
flask-cors==4.0.1
